Improved readme, added "dozens" and "high/low" feature and improved code for roulette_cog
The readme now has some much needed imrpovements: included an intro to the bot, fixed the commands list and added a guide to help get the bot running.
Roulette_cog now has the option to bet on dozens and on high/low and the overall code was imrpoved, particularly new error messages were added to catch undesired situations like users not specifying bets.